Include concept of acceleration to process
Now particles have some inertia to mantain its own trajectory.
This is important when all random initial solutions are located in a cluster far from the answer.
In this case, particles won't collide to an mutual answer in between them, but instead they will orbit the mean point and look for options around it.